refactor: consolidate scattered utility code into lib/
Move non-component logic out of components/ and app/ into lib/ with a
consistent convention: index.ts (universal), server.ts (server-only),
action.ts (server actions).

- lib/product/index.ts: variant selection + optimistic cart info
- lib/cart/action.ts + server.ts: cart actions + cache invalidation
- lib/search/action.ts: predictive search action
- lib/collections/server.ts: collection data fetching
- lib/utils/index.ts: consolidated all sub-module utilities
- Inline buildProductMetadata into page.tsx, DEBOUNCE_MS into context.tsx
- Fix backwards dependency (lib/shopify/transforms → components/pdp/variants)
