+++ apps/template/app/products/[handle]/page.tsx
@@ -3,10 +3,8 @@ import { notFound } from "next/navigation";
 
 import { ProductDetailPage } from "@/components/pdp/product-detail-page";
 import { getLocale } from "@/lib/params";
-
+import { buildAlternates, buildOpenGraph } from "@/lib/seo";
 import { getProduct } from "@/lib/shopify/operations/products";
-
-import { buildProductMetadata } from "./shared";
 
 const PLACEHOLDER_HANDLE = "__placeholder__";
 
@@ -22,6 +20,45 @@ export async function generateMetadata({
   if (handle === PLACEHOLDER_HANDLE) return {};
 
   return buildProductMetadata(handle, locale, `/products/${handle}`);
+}
+
+async function buildProductMetadata(
+  handle: string,
+  locale: string,
+  canonicalPath: string,
+): Promise<Metadata> {
+  const product = await getProduct(handle, locale).catch(() => notFound());
+  const images = product.featuredImage
+    ? [
+        {
+          url: product.featuredImage.url,
+          width: product.featuredImage.width,
+          height: product.featuredImage.height,
+          alt: product.featuredImage.altText,
+        },
+      ]
+    : ["/og-default.png"];
+
+  return {
+    title: product.seo.title,
+    description: product.seo.description,
+    alternates: buildAlternates({
+      pathname: canonicalPath,
+    }),
+    openGraph: buildOpenGraph({
+      title: product.seo.title,
+      description: product.seo.description,
+      url: canonicalPath,
+      type: "website",
+      images,
+    }),
+    twitter: {
+      card: "summary_large_image",
+      title: product.seo.title,
+      description: product.seo.description,
+      images,
+    },
+  };
 }
 
 export const unstable_instant = {
