--- apps/template/components/cart/constants.ts
@@ -1,6 +1,0 @@
-/**
- * Debounce delay for cart quantity updates.
- * First click fires immediately (leading edge), subsequent clicks within
- * this window are batched and sent after the delay (trailing edge).
- */
-export const DEBOUNCE_MS = 400;
