+++ apps/template/components/pdp/option-picker.tsx
@@ -4,8 +4,7 @@ import type { ComponentPropsWithoutRef } from "react";
 import type { ProductOption, ProductVariant } from "@/lib/types";
 import { cn } from "@/lib/utils";
 
-import type { SelectedOptions } from "./variants";
-import { getVariantUrl } from "./variants";
+import { type SelectedOptions, getVariantUrl } from "@/lib/product";
 
 interface OptionPickerProps extends ComponentPropsWithoutRef<"div"> {
   option: ProductOption;
