+++ apps/template/components/product-card.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { getTranslations } from "next-intl/server";
 
-import { productCardToOptimisticInfo } from "@/components/cart/optimistic-info";
+import { productCardToOptimisticInfo } from "@/lib/product";
 import {
   ProductCardBadge,
   ProductCardContent,
