+++ apps/template/components/search/results.tsx
@@ -12,7 +12,7 @@ import { Skeleton } from "@/components/ui/skeleton";
 import type { Locale } from "@/lib/i18n";
 import { buildProductFiltersFromParams, getProducts } from "@/lib/shopify/operations/products";
 import { transformShopifyFilters } from "@/lib/shopify/transforms/filters";
-import { RESULTS_PER_PAGE } from "@/lib/utils/product-card";
+import { RESULTS_PER_PAGE } from "@/lib/utils";
 
 const RESULTS_SKELETON_KEYS = Array.from(
   { length: 10 },
