--- apps/template/lib/utils/format.ts
@@ -1,7 +1,0 @@
-export function formatPrice(amount: number, currencyCode: string, locale: string): string {
-  return new Intl.NumberFormat(locale, {
-    style: "currency",
-    currency: currencyCode,
-    currencyDisplay: "narrowSymbol",
-  }).format(amount);
-}
