+++ apps/template/lib/utils/index.ts
@@ -1,2 +1,49 @@
-export { cn } from "./cn";
-export { formatPrice } from "./format";
+import { type ClassValue, clsx } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
+
+export function formatPrice(amount: number, currencyCode: string, locale: string): string {
+  return new Intl.NumberFormat(locale, {
+    style: "currency",
+    currency: currencyCode,
+    currencyDisplay: "narrowSymbol",
+  }).format(amount);
+}
+
+export function parseFiltersFromSearchParams(
+  searchParams: Record<string, string | string[] | undefined>,
+): Record<string, string | string[] | undefined> {
+  const filters: Record<string, string | string[] | undefined> = {};
+
+  for (const [key, value] of Object.entries(searchParams)) {
+    if (!key.startsWith("filter.") || value === undefined) {
+      continue;
+    }
+    filters[key] = value;
+  }
+
+  return filters;
+}
+
+export function searchParamsToRecord(
+  searchParams: URLSearchParams,
+): Record<string, string | string[] | undefined> {
+  const record: Record<string, string | string[] | undefined> = {};
+
+  for (const key of new Set(searchParams.keys())) {
+    const values = searchParams.getAll(key);
+
+    if (values.length === 0) {
+      continue;
+    }
+
+    record[key] = values.length === 1 ? values[0] : values;
+  }
+
+  return record;
+}
+
+export const RESULTS_PER_PAGE = 50;
