--- apps/template/lib/utils/product-card.ts
@@ -1,1 +1,0 @@
-export const RESULTS_PER_PAGE = 50;
